backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
transformers==4.40.0
torch==2.3.0
sentencepiece==0.2.0
pydantic==2.7.0
pandas==2.2.0
python-multipart==0.0.9
python-dotenv==1.0.0

